Refactored Code to Prevent Memory Leak on Network Request
FollowersListVC makes a network request via NetworkManager to get a list of followers for the CollectionView when a username is searched. It stores that list of followers using a self reference in the success case which can cause a memory leak due to a strong reference link to the FollowersListVC. In order to prevent that, set [weak self] for the result that way, when the view is removed it will no longer hold onto that strong reference and can be removed from memory.
