+++ GithubUsers/GithubUsers/Utilities/UIHelper.swift
@@ -1,0 +1,27 @@
+//
+//  UIHelper.swift
+//  GithubUsers
+//
+//  Created by Marco Capraro on 2/27/24.
+//
+
+import UIKit
+
+struct UIHelper {
+    
+    static func createThreeColumnFlowLayout(in view: UIView) -> UICollectionViewFlowLayout {
+        // 3 Column Flow Layout used for our UICOllectionView
+        let width = view.bounds.width
+        let padding: CGFloat = 12
+        let minimumSpacing: CGFloat = 10
+        let availableWidth = width - (padding * 2) - (minimumSpacing * 2)
+        let itemWidth = availableWidth / 3
+        
+        let flowLayout = UICollectionViewFlowLayout()
+        flowLayout.sectionInset = UIEdgeInsets(top: padding, left: padding, bottom: padding, right: padding)
+        flowLayout.itemSize = CGSize(width: itemWidth, height: itemWidth + 40)
+        
+        return flowLayout
+    }
+    
+}
